fix(Filter): Filter getting unmounted losing filter state
The filter gets unmounted when the dropdown is closed. Previous fix to make the filter a controlled
component had no effect because of this. I moved the filterTerm to the Picky component but the
performance hit has too huge on large lists, causing a full tree render on input.

fix #73